Moving SpecialCases into separate module Method to increment episode numbers based on amount that's shown today
